# doc/api.yaml
openapi: 3.0.1
info:
  title: WASAPhoto APIs
  version: 1.0.3
  description: >
    **"Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!"**
    
    WASAPhoto is web application based on RESTfull architecture to allows upload your photos directly from your PC, and they will be visible to everyone following you.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  
  contact:
    name: Andrea Pizzi
    url: https://github.com/pizzi-andrea/WASAPhoto
    email: pizzi.1995517@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
paths:
  /users/:
    summary: users
    description: >
      this resource rappresent collection of users
    
    get:
      security:
        - BearerAuth: []
        
      tags:
        - user
      summary: list registred users
      description: >
        this method return the list of users currently registered
      parameters: 
        
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
          
          
      operationId: listUsers
      
      responses:
        "200":
          description: >
            get the list contaning all users current signed
          content:
            application/json:
              schema:
                description: >
                  list of users  
                type: array
                items: 
                  $ref: "#/components/schemas/User"
                minItems: 1
                maxItems: 300
             
                example:
                  - username: "MarioRoss"
                    uid: 10
                  - username: "alux"
                    uid: 840
        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{uid}/:
    summary:  users
    description: >
       Resource rappresent a single user
       
    parameters:
        - $ref: "#/components/parameters/Uid"
    
    get:
      tags:
        - user
      summary: get specific user profile
      description: >
        give user id and put user profile. User profile is object contain 
        all information on user, in particular:
          - information about user
          - stream photos updated
          - number of photo have been updated
          - number of followers
          - number of following
      
      security:
        - BearerAuth: []
        
      operationId: getUserProfile
      
      responses:
        "200":
          description: >
            get user profile corresponding to *uid*
          content:
            application/json:
              schema:
                {$ref: "#/components/schemas/Profile"}
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
        "400":
          $ref: "#/components/responses/BadRequest"

    put:
      tags:
        - user
      summary: assign or update username
      description: >
        assign username or update old username with new username.
        The username to set is in the body request
      
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
        description: >
          new username
        required: true
      
      operationId: setMyUserName
      
      security:
        - BearerAuth: []
      
      responses:
        "204":
          description: >
            success, assigned new username
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "404":
          description: >
            $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{uid}/followers/:
    summary: user followers
    description: >
      this resurce rappresent the followers of specific user
    parameters:
      - $ref: "#/components/parameters/Uid"
    get:
      security:
        - BearerAuth: []
      tags:
        - user
        - follower
      summary: get all followers
      description: >
        give a UID return a list contanings all followers user
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
          
      operationId: listFollowers
      
      responses:
        "200":
          description: >
            success, return list of followers
          content:
            application/json:
              schema:
                description: >
                  list of followers
                type: array
                items: 
                  $ref: "#/components/schemas/User"
                minItems: 1
                maxItems: 1000
        "404":
          $ref: "#/components/responses/UserNotFound"
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/banned/:
    summary: banned users
    description: >
      this resource rappresent user banned by uid user
    
    parameters:
        - $ref: "#/components/parameters/Uid"
    
    get:
      security:
        - BearerAuth: []
        
      tags:
        - user
        - ban
      summary: users banned
      description: >
        given *uid* then list all user banned by user associated at *uid* 
    
      parameters:

        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
      
      operationId: listBannedUsers
      
      responses:
        "200":
          description: >
            success, getted all user banned
          
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
                minItems: 1
                maxItems: 100
                description: >
                  list  of banned users

        "404":
          $ref: "#/components/responses/UserNotFound"
        
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/: 
    
    parameters:
      - $ref: "#/components/parameters/Uid"
    
    post:
      security:
        - BearerAuth: []
        
      tags:
        - photo
      summary: update photo
      description: >
        give photo and update it
        
      requestBody:
        description: >
          photo to update
        content:
          multipart/form-data:
            schema: 
              $ref: "#/components/schemas/Photo"
      
      operationId: uploadPhoto
      
      responses:
        "201":
          description: >
            success, update photo
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Photo"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      security:
        - BearerAuth: []
      tags:
        - photo
      summary: list stream photos updated
      description: >
        given *uid* get all photo has updated
      
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - name: sortBy
          in: query
          required: false
          schema:
            description: >
              options for sorting photos
            type: string
            enum: ["dataUpdate"]   # future updates [, "size", "likes", "comments"]
            default: "dataUpdate"
          description: >
            ordering photo getted by value in enum
        - name: ordering
          in: query
          required: false
          schema:
            description: >
              flag for reverse sorting
            type: string
            enum: ["asc", "desc"]
            default: "desc"
          description: >
            order photo in reverse order
      
      operationId: listPhoto
      
      responses:
        "200":
          description: >
            success, stream photos updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamPhotos"
        
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    QueryLimit:
      description: >
        security parameter to limit the number of items in a response and avoid DOS attacks
      name: limit
      in: query
      required: false
      schema:
        description: >
          format limit values
        type: integer
        format: int
        minimum: 1
        example: 50
    UsernameFilter:
      description: >
        research in list of users by username 
      name: username
      in: query
      required: false
      schema:
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 16
        example: "marioross"
    Uid:
      name: uid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user id
...
  responses:
    UnauthorizedError:
      description: >
        Access token is missing or invalid 
    UnauthorizedToken:
      description: >
        the submitted token does not have the necessary permissions to access this resource
    ServerError:
      description: >
        server internal error, for more information see log file
    UserNotFound:
      description: >
        Not found, user not found
    BadRequest:
      description: >
       Bad request, formating error 
  securitySchemes:
    
    BearerAuth:
      description: >
        basic safety mechanism based on "bearer" mechanism. APIs that require a minimum level of security can adopt this method
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      description: >
        this object rappresent a user
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/Id"
        username: 
          $ref: "#/components/schemas/Username"
      required:
        - uid
        - username
...
    Username:
      description: >
        username of a user
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: "marioross"
    Id:
      description: >
        Identificator at 64-bit
      type: integer
      format: int64 
      readOnly: true
      example: 4570 
    Profile:
      description: >
        user profile rappresentation
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        stream:
          $ref: "#/components/schemas/StreamPhotos"
        follower:
          type: integer
          minimum: 0
          description: >
            number user that follow a specific user
          example: 12
        following:
          type: integer
          minimum: 0
          description: >
            numer of users following by specific user
          example: 70
    Photo:
      description: >
        this object rappresent a photo
      type: object
      properties:
        photoId:
          $ref: "#/components/schemas/Id"
        timeUpdate:
          $ref: "#/components/schemas/TimeStamp"

        imageData:
          description: >
            data
          type: string
          format: binary
          minLength: 0
          maxLength: 5000000
          

        
        descriptionImg:
          description: >
            image description
          type: string
          example: "a good photo!"
          minLength: 1
          maxLength: 250
          pattern: "^.*?$"
      required:
        - imageData
        - timeUpdate
    StreamPhotos:
      description: >
        model of stream of photos
      type: array
      items:
        $ref: "#/components/schemas/Photo"
      minItems: 0
      maxItems: 100
    TimeStamp:
      description: >
        this components describe timestamp value
        conform to [RFC3339](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6) specification
      type: string
      format: data-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 21
      readOnly: true
...